all services communicate through ocelot

+++ TwitterMock/CommentService/Controllers/CommentController.cs
@@ -9,6 +9,7 @@ namespace CommentService.Controllers;
 public class CommentController : ControllerBase
 {
     private readonly ICommentService _commentService;
+    private readonly HttpClient _client = new HttpClient();
 
     public CommentController(ICommentService commentService)
     {
@@ -17,10 +18,15 @@ public class CommentController : ControllerBase
 
     [HttpGet]
     [Route("{postId}")]
-    public async Task<IActionResult> GetComments([FromRoute] int postId, [FromQuery] PaginatedDto dto)
+    public async Task<IActionResult> GetComments([FromRoute] int postId, [FromQuery] PaginatedDto dto, [FromHeader] string token)
     {
         try
         {
+            var url = "http://localhost:5206/api/Auth/ValidateToken";
+            _client.DefaultRequestHeaders.Add("token", token);
+            var result = await _client.GetAsync(url);
+
+            if (!result.IsSuccessStatusCode) return Unauthorized(result.RequestMessage);
             return Ok(await _commentService.GetComments(postId, dto));
         }
         catch (Exception e)
@@ -30,10 +36,16 @@ public class CommentController : ControllerBase
     }
 
     [HttpPost]
-    public async Task<IActionResult> AddComment([FromBody] AddCommentDto dto)
+    public async Task<IActionResult> AddComment([FromBody] AddCommentDto dto, [FromHeader] string token)
     {
         try
         {
+            var url = "http://localhost:5206/api/Auth/ValidateToken";
+            _client.DefaultRequestHeaders.Add("token", token);
+            var result = await _client.GetAsync(url);
+
+            if (!result.IsSuccessStatusCode) return Unauthorized(result.RequestMessage);
+            
             await _commentService.AddComment(dto);
             return StatusCode(201, "Comment successfully added");
         }
@@ -45,10 +57,16 @@ public class CommentController : ControllerBase
 
     [HttpPut]
     [Route("{commentId}")]
-    public async Task<IActionResult> UpdateComment([FromRoute] int commentId, [FromBody] UpdateCommentDto dto)
+    public async Task<IActionResult> UpdateComment([FromRoute] int commentId, [FromBody] UpdateCommentDto dto, [FromHeader] string token)
     {
         try
         {
+            var url = "http://localhost:5206/api/Auth/ValidateToken";
+            _client.DefaultRequestHeaders.Add("token", token);
+            var result = await _client.GetAsync(url);
+
+            if (!result.IsSuccessStatusCode) return Unauthorized(result.RequestMessage);
+            
             await _commentService.UpdateComment(commentId, dto);
             return Ok();
         }
@@ -60,10 +78,16 @@ public class CommentController : ControllerBase
 
     [HttpDelete]
     [Route("{commentId}")]
-    public async Task<IActionResult> DeleteComment([FromRoute] int commentId)
+    public async Task<IActionResult> DeleteComment([FromRoute] int commentId, [FromHeader] string token)
     {
         try
         {
+            var url = "http://localhost:5206/api/Auth/ValidateToken";
+            _client.DefaultRequestHeaders.Add("token", token);
+            var result = await _client.GetAsync(url);
+
+            if (!result.IsSuccessStatusCode) return Unauthorized(result.RequestMessage);
+            
             await _commentService.DeleteComment(commentId);
             return Ok();
         }
